feat: add Pagination for Loan Service and tests

+++ infrastructure/entry-points/reactive-web/src/main/java/com/creditya/loanservice/api/dto/response/LoanResponseDTO.java
@@ -16,6 +16,6 @@ public record LoanResponseDTO(
 
         BigDecimal interestRate,
         BigDecimal totalMontlyDebt,
-        Integer approvedLoans
+        Long approvedLoans
 ) {}
